team8/ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
Pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0

